Refactor(gahee/circle-home) : 동아리 홈 폴더명들 변경 circle-home => home / CircleHome => Home 및 폴더명 변경에 따른 app.js, home.ctrl.js 코드 변경 #428
